📈 Add clients metrics collection via the analysis server
