+++ plugins/analysis/server/events.go
@@ -21,6 +21,8 @@ var Events = struct {
 	Heartbeat *events.Event
 	// FPCHeartbeat triggers when an FPC heartbeat has been received.
 	FPCHeartbeat *events.Event
+	// MetricHeartbeat triggers when an MetricHeartbeat heartbeat has been received.
+	MetricHeartbeat *events.Event
 }{
 	events.NewEvent(stringCaller),
 	events.NewEvent(stringCaller),
@@ -29,6 +31,7 @@ var Events = struct {
 	events.NewEvent(errorCaller),
 	events.NewEvent(heartbeatPacketCaller),
 	events.NewEvent(fpcHeartbeatPacketCaller),
+	events.NewEvent(metricHeartbeatPacketCaller),
 }
 
 func stringCaller(handler interface{}, params ...interface{}) {
@@ -49,4 +52,8 @@ func heartbeatPacketCaller(handler interface{}, params ...interface{}) {
 
 func fpcHeartbeatPacketCaller(handler interface{}, params ...interface{}) {
 	handler.(func(hb *packet.FPCHeartbeat))(params[0].(*packet.FPCHeartbeat))
+}
+
+func metricHeartbeatPacketCaller(handler interface{}, params ...interface{}) {
+	handler.(func(hb *packet.MetricHeartbeat))(params[0].(*packet.MetricHeartbeat))
 }
