+++ plugins/metrics/parameters.go
@@ -1,6 +1,10 @@
 package metrics
 
-import "time"
+import (
+	"time"
+
+	flag "github.com/spf13/pflag"
+)
 
 const (
 	// should always be 1 second
@@ -13,3 +17,15 @@ const (
 	MemUsageMeasurementInterval    = 1 * time.Second
 	SyncedMeasurementInterval      = 1 * time.Second
 )
+
+const (
+	// CfgMetricsLocal defines the config flag to enable/disable local metrics.
+	CfgMetricsLocal = "metrics.local"
+	// CfgMetricsGlobal defines the config flag to enable/disable global metrics.
+	CfgMetricsGlobal = "metrics.global"
+)
+
+func init() {
+	flag.Bool(CfgMetricsLocal, true, "include local metrics")
+	flag.Bool(CfgMetricsGlobal, false, "include global metrics")
+}
