+++ plugins/prometheus/analysis_server.go
@@ -1,0 +1,1 @@
+package prometheus
